Refactor PartAttachmentWidget
- Now displays generic attachments
- Augment the InvenTreeAttachment class to "know" about its model reference
