+++ lib/widget/part_detail.dart
@@ -6,7 +6,7 @@ import "package:inventree/app_colors.dart";
 import "package:inventree/inventree/stock.dart";
 import "package:inventree/l10.dart";
 import "package:inventree/helpers.dart";
-import "package:inventree/widget/part_attachments_widget.dart";
+import "package:inventree/widget/attachment_widget.dart";
 import "package:inventree/widget/part_notes.dart";
 import "package:inventree/widget/progress.dart";
 import "package:inventree/inventree/part.dart";
@@ -410,7 +410,10 @@ class _PartDisplayState extends RefreshableState<PartDetailWidget> {
           Navigator.push(
             context,
             MaterialPageRoute(
-              builder: (context) => PartAttachmentsWidget(part)
+              builder: (context) => AttachmentWidget(
+                  InvenTreePartAttachment(),
+                  part.pk,
+                  InvenTreeAPI().checkPermission("part", "change"))
             )
           );
         },
